requirements.txt
crewai[google-genai]
crewai-tools
python-dotenv
streamlit
duckduckgo-search
langchain_google_genai
langchain-community
langchain-groq
